requirements.txt
PySide6
PySerial